requirements.txt
streamlit==1.32.0
openai==1.12.0
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.3
plotly==5.18.0
python-dotenv==1.0.0
SQLAlchemy==2.0.25
joblib==1.3.2
pytest==7.4.4
black==24.1.1
flake8==7.0.0 